requirements.txt
sentence-transformers
PyMuPDF
python-docx
python-pptx
pandas
torch
